requirements.txt
fastapi
uvicorn
langchain
langchain-community
huggingface-hub
chromadb
sentence-transformers
psycopg2-binary
pandas
python-dotenv
tqdm
langchain-core
pydantic
Ollama